Move server integration test to "normal" task
as it no longer requires the coroutine client flag to be set.
