+++ thrifty-integration-tests/build.gradle
@@ -77,6 +77,7 @@ def kompileTestThrift = tasks.register("kompileTestThrift", JavaExec) { t ->
             "--map-type=java.util.LinkedHashMap",
             "--set-type=java.util.LinkedHashSet",
             "--list-type=java.util.ArrayList",
+            "--experimental-kt-generate-server",
             "$projectDir/ClientThriftTest.thrift"
     ]
 }
@@ -92,7 +93,6 @@ def kompileCoroutineTestThrift = tasks.register("kompileCoroutineTestThrift", Ja
             "--kt-file-per-type",
             "--kt-emit-jvmname",
             "--service-type=coroutine",
-            "--experimental-kt-generate-server",
             "$projectDir/CoroutineClientTest.thrift"
     ]
 }
